Refactor unrecognised job
This PR refactors the unrecognised job so that it sits within
the IamRemediationService, which includes the outdated
credentials job. This is better, because both the jobs are
part of the same idea that Security HQ should automate IAM
vulnerabilities. These two jobs now live together in
IamRemediationService.scala.
