+++ hq/app/logic/IamOutdatedCredentials.scala
@@ -11,22 +11,7 @@ import utils.attempt.{Attempt, FailedAttempt, Failure}
 import scala.concurrent.ExecutionContext
 
 
-object IamRemediation extends Logging {
-  /**
-    * Removes the FailedAttempts from the Either and returns a list of tuples with only the Right values.
-    * This function uses generics to make it easier to test, but to avoid confusion it was written to take
-    * a Map of AWSAccount to Either and return a list of tuples of AWS Account to CredentialReportDisplay.
-    */
-  def getCredsReportDisplayForAccount[A, B](allCreds: Map[A, Either[FailedAttempt, B]]): List[(A, B)] = {
-    allCreds.toList.foldLeft[List[(A, B)]](Nil) {
-      case (acc, (_, Left(failure))) =>
-        logger.error(s"unable to generate credential report display: ${failure.logMessage}")
-        acc
-      case (acc, (account, Right(credReportDisplay))) =>
-        (account, credReportDisplay) :: acc
-    }
-  }
-
+object IamOutdatedCredentials extends Logging {
   /**
     * Look through all credentials reports to find users with expired credentials,
     * see below for more detail (`identifyUsersWithOutdatedCredentials`).
